+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Changed
+* Document Component: Kenmerken en kaart paneel - Heading structuur kloppend maken ([#3636](https://github.com/dso-toolkit/dso-toolkit/issues/3636))
+
 ### Fixed
 * Info Button: WCAG - toggletip content wordt niet voorgelezen bij openen van de tooltip ([#3638](https://github.com/dso-toolkit/dso-toolkit/issues/3638))